Created new make announcement page
