+++ frontend/src/App.js
@@ -9,6 +9,7 @@ import Calendar from './components/calendar/Calendar';
 import Navbar from "./components/Navbar";
 import StandardPage from "./components/standard-page/StandardPage";
 import AnnouncementPage from "./components/announcements/AnnouncementPage";
+import CreateAnnouncementPage from "./components/announcements/CreateAnnouncementPage";
 import GoogleSignIn from "./components/Login/GoogleSignIn";
 import Store from './store/Store';
 
@@ -22,6 +23,7 @@ function App() {
             <Route path="/" exact></Route>
             <Route path="/login"><Login /></Route>
             <Route path="/announcements"> <AnnouncementPage> </AnnouncementPage></Route>
+            <Route path="/create-announcements"> <CreateAnnouncementPage> </CreateAnnouncementPage></Route>
             <Route path="/login/:type"></Route>
             <Route path="/calendar"> {/* Temporary route for now */}
               <Calendar />
