+++ frontend/src/components/announcements/CreateAnnouncementPage.js
@@ -1,0 +1,18 @@
+import React from "react";
+import CreateAnnouncementBlock from './CreateAnnouncementBlock';
+
+class CreateAnnouncementPage extends React.Component {
+    constructor(props) {
+        super(props);
+    }
+	render(){
+        
+		return(
+            <div>
+                <CreateAnnouncementBlock />
+            </div>
+		);
+	}
+}
+
+export default CreateAnnouncementBlock
